admin/requirements.txt
Django==5.1.4
djangorestframework==3.14.0
mysqlclient==2.2.7
django-mysql==4.16.0
django-cors-headers==3.5.0
pika==1.3.2
pillow
